requirements.txt
# Web Server
fastapi
uvicorn
websockets>=11.0.3

# Google Gemini
google-generativeai==0.5.4
jsonschema
supabase-async

# Database
supabase>=2.0.0

# Utilities
python-dotenv

# Python version compatibility
taskgroup
exceptiongroup

# For Railway deployment, we don't need these client-side dependencies
# opencv-python
# pyaudio
# pillow
# mss

# Note: audioop is a built-in Python module and doesn't need to be installed